+++ pkg/dartdev/test/native_assets/helpers.dart
@@ -114,7 +114,8 @@ Future<void> copyTestProjects(Uri copyTargetUri, Logger logger,
       'code_assets',
       'data_assets',
       'hooks',
-      'native_toolchain_c'
+      'native_toolchain_c',
+      'record_use',
     ])
       package: {
         'path': sdkRoot
@@ -123,9 +124,6 @@ Future<void> copyTestProjects(Uri copyTargetUri, Logger logger,
       },
     'meta': {
       'path': sdkRoot.resolve('pkg/meta/').toFilePath(),
-    },
-    'record_use': {
-      'path': sdkRoot.resolve('pkg/record_use/').toFilePath(),
     },
   };
   final userDefinesWorkspace = {};
@@ -285,7 +283,7 @@ Future<void> recordUseTest(
       packageUnderTest,
       fun,
       const ['drop_dylib_recording', 'drop_data_asset'],
-      sdkRootUri.resolve('pkg/record_use/'),
+      sdkRootUri.resolve('third_party/pkg/native/pkgs/record_use/'),
       sdkRootUri,
       false,
     );