+++ pkg/linter/lib/src/rules/always_require_non_null_named_parameters.dart
@@ -4,21 +4,13 @@
 
 import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_state.dart';
-import 'package:analyzer/error/error.dart';
 
 import '../analyzer.dart';
-import '../diagnostic.dart' as diag;
 
 const _desc = r'Specify `@required` on named parameters without defaults.';
 
-class AlwaysRequireNonNullNamedParameters extends AnalysisRule {
-  AlwaysRequireNonNullNamedParameters()
-    : super(
-        name: LintNames.always_require_non_null_named_parameters,
-        description: _desc,
-        state: RuleState.removed(since: dart3_3),
-      );
-
-  @override
-  DiagnosticCode get diagnosticCode => diag.removedLint;
-}
+final alwaysRequireNonNullNamedParameters = RemovedAnalysisRule(
+  name: LintNames.always_require_non_null_named_parameters,
+  description: _desc,
+  since: dart3_3,
+);