+++ pkg/linter/lib/src/rules/invariant_booleans.dart
@@ -4,22 +4,14 @@
 
 import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_state.dart';
-import 'package:analyzer/error/error.dart';
 
 import '../analyzer.dart';
-import '../diagnostic.dart' as diag;
 
 const _desc =
     r'Conditions should not unconditionally evaluate to `true` or to `false`.';
 
-class InvariantBooleans extends AnalysisRule {
-  InvariantBooleans()
-    : super(
-        name: LintNames.invariant_booleans,
-        description: _desc,
-        state: RuleState.removed(since: dart3),
-      );
-
-  @override
-  DiagnosticCode get diagnosticCode => diag.removedLint;
-}
+final invariantBooleans = RemovedAnalysisRule(
+  name: LintNames.invariant_booleans,
+  description: _desc,
+  since: dart3,
+);