+++ pkg/linter/lib/src/rules/list_remove_unrelated_type.dart
@@ -4,21 +4,13 @@
 
 import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_state.dart';
-import 'package:analyzer/error/error.dart';
 
 import '../analyzer.dart';
-import '../diagnostic.dart' as diag;
 
 const _desc = r'Invocation of `remove` with references of unrelated types.';
 
-class ListRemoveUnrelatedType extends AnalysisRule {
-  ListRemoveUnrelatedType()
-    : super(
-        name: LintNames.list_remove_unrelated_type,
-        description: _desc,
-        state: RuleState.removed(since: dart3_3),
-      );
-
-  @override
-  DiagnosticCode get diagnosticCode => diag.removedLint;
-}
+final listRemoveUnrelatedType = RemovedAnalysisRule(
+  name: LintNames.list_remove_unrelated_type,
+  description: _desc,
+  since: dart3_3,
+);