+++ pkg/linter/lib/src/rules/prefer_equal_for_default_values.dart
@@ -4,21 +4,13 @@
 
 import 'package:analyzer/analysis_rule/analysis_rule.dart';
 import 'package:analyzer/analysis_rule/rule_state.dart';
-import 'package:analyzer/error/error.dart';
 
 import '../analyzer.dart';
-import '../diagnostic.dart' as diag;
 
 const _desc = r'Use `=` to separate a named parameter from its default value.';
 
-class PreferEqualForDefaultValues extends AnalysisRule {
-  PreferEqualForDefaultValues()
-    : super(
-        name: LintNames.prefer_equal_for_default_values,
-        description: _desc,
-        state: RuleState.removed(since: dart3),
-      );
-
-  @override
-  DiagnosticCode get diagnosticCode => diag.removedLint;
-}
+final preferEqualForDefaultValues = RemovedAnalysisRule(
+  name: LintNames.prefer_equal_for_default_values,
+  description: _desc,
+  since: dart3,
+);